Got openCV video stream to work
Simplified some of the openCV code so that we don't get those errors anymore. You can see the video run on <device_URL>/camera/video_feed. Note the changes to opencv's command line; using 16 frames in the argument seems to decrease CPU utilization, which we'll have to look into in the future.